refactor(backend): create services folder in users module
