+++ packages/backend/src/modules/users/index.ts
@@ -1,7 +1,4 @@
 export { UsersModule } from './UsersModule';
 export * from './usersExceptions';
-export { UsersService } from './UsersService';
-export * from './CreateUserCommand';
-export * from './UpdateUserCommand';
-export * from './GetUserQuery';
+export * from './services';
 export * from './ports';
