+++ packages/backend/src/modules/users/services/index.ts
@@ -1,0 +1,4 @@
+export * from './CreateUserCommand';
+export * from './GetUserQuery';
+export * from './UpdateUserCommand';
+export * from './UsersService';
